+++ src/commands/mod.rs
@@ -1,6 +1,8 @@
+pub mod attach;
 pub mod clean;
 pub mod clean_all;
 pub mod config;
+pub mod exec;
 pub mod info;
 pub mod list;
 pub mod logs;